Changed Layer for Cuttable
The cuttable layer was set to 9, but due to the recent merge, it had to be moved to layer 10. This commit sets the sliceable layer to layer 10 and changes all the code that uses the sliceable layer. The code now has 1 variable per class, which controls the sliceable layer, rather than a hardcoded '10' every time the layer is needed.