Separated Blending and Cutting
Along with separating blending and cutting, I also made it so that the level of liquid goes up as something is blended (need some creative ways to vary the amount of liquid added to the blend, but that is a later problem). On top of those changes, made some minor improvements to blending and slicing scripts.